[FIX][REFACTOR] fixed track lengthKm search precision - Replaced Double/real with BigDecimal/numeric(10,3) to preserve exact decimal precision when searching by track length. Updated affected DTOs, mappings, and controller tests.

+++ src/main/java/sk/mkrajcovic/challenges/controller/dto/TrackDetailResponse.java
@@ -1,10 +1,12 @@
 package sk.mkrajcovic.challenges.controller.dto;
 
+import java.math.BigDecimal;
+
 public record TrackDetailResponse(
 
 	Integer id,
 	String country,
 	String name,
-	Double lengthKm
+	BigDecimal lengthKm
 
 ) { }
